requirements.txt
flask==2.2.3
tensorflow==2.12.0
pillow==9.4.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
werkzeug==2.2.3